Calculate ideal stagger size and nonces by default
Instead of forcing the user to pre-calculate the appropriate nonces and stagger
sizes for optimized execution, attempt to automatically calculate everything
the user hasn't specified.

Optimized nonces per thread are now required for SSE4 as well, to simplify the
code and get rid of the inefficient fallback-calls and corresponding warnings.